+++ app/assets/config/manifest.js
@@ -1,2 +1,5 @@
 //= link_tree ../images
-//= link_directory ../stylesheets .css
+//= link application.css
+//= link_tree ../../javascript .js
+//= link_tree ../../../vendor/javascript .js
+//= link rails-ujs.js
