+++ app/javascript/application.js
@@ -1,0 +1,11 @@
+
+document.addEventListener("DOMContentLoaded", function () {
+  const hamburger = document.getElementById("hamburger-btn");
+  const nav = document.getElementById("header-nav");
+
+  if (hamburger && nav) {
+    hamburger.addEventListener("click", function () {
+      nav.classList.toggle("open");
+    });
+  }
+});
