+++ config/importmap.rb
@@ -1,0 +1,8 @@
+# Pin npm packages by running ./bin/importmap
+pin "application", preload: true
+pin "@hotwired/turbo-rails", to: "turbo.min.js", preload: true
+pin "@hotwired/stimulus", to: "stimulus.min.js", preload: true
+pin "@hotwired/stimulus-loading", to: "stimulus-loading.js", preload: true
+pin "@rails/activestorage", to: "activestorage.esm.js", preload: true
+pin_all_from "app/javascript/controllers", under: "controllers"
+pin "@rails/ujs", to: "rails-ujs.js"
